refactor: Refactor admin portal for campaign managers functionality
This commit refactors the campaign managers functionality in admin portal, and includes handling addition as well as removal of managers.
Importantly, the handling of removal is shifted from the CampaignManagersView component to the Managers component, enhancing separation of concerns.
